Create README.md_Mister Passportu
Teaching the model the basic skills of generalization, analysis and generation of qualitative reasoning based on the available information in such areas as: logic, philosophy, critical thinking, the strategy of cooperation and mercy for sapiens, humor, the fundamentals of the philosophy of culture and design. Too ambitious, I know. But as I see it, these are good goals. Culture is like a pie that only grows with the number of its bearers, let's do useful and accessible things.